refactor(stories): host the chat transcript in the real shell
Product/Chat Surface mounted ChatView and Composer inside a 960x760 centred
card with a border and a radius — a frame the product does not have. In the
app the transcript lives in the full-height maka-panel-detail pane that
app-shell.stories.tsx already composes from the real AstryxAppShell, so every
geometry comparison against the old story was measuring its own scaffolding.
Two of its six stories were also the shell stories' states in that wrong
frame: StreamingResponse against StreamingTurn, and the two empty chats
against EmptyHome, which is the only empty home the app renders.

Move the states the shell did not have — the reasoning-and-tool-call
transcript and the session context layer — onto ComposedShell, and delete the
parallel file with its second copy of the session, message and composer
fixtures. Widening the session overlay to Partial<SessionSummary> keeps the
one-active-session rule while letting a story set the name and labels the
context layer reads.

SessionContextNarrow is not carried over: it faked a narrow column with a
360px div, and the manifest now renders the same story at compact and floor.
